requirements.txt
dj-database-url==0.5.0
Django==2.2
Django==2.2
django-adminlte2==0.4.1
django-adminlte2==0.4.1
django-heroku==0.3.1
django-heroku==0.3.1
djangorestframework==3.9.3
gunicorn==19.9.0
gunicorn==19.9.0
pytz==2019.1
pytz==2019.1
sqlparse==0.3.0
sqlparse==0.3.0
